+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.2] / 2023-04-17
+### UI
+- Update `NewPushButtonData` the default `Text` equals to `targetName`.
+### Tests
+- Add `RevitCreateItemsCommandTests`
+- Add `RevitCreateItemsCommandSetTests`
+- Update `RevitCreateItemsWithNameTests`
+
 ## [0.5.1] / 2023-04-14
 ### RibbonStackExtension
 - Add RowStackedItems